Replace the SPI_WAIT_FOR_AR macro with function
Replace the SPI_WAIT_FOR_AR macro with the function, which
does the same but doesn't modify variables out of its score
(the macro did). I think there is no need to declare the
function as inline there, because it isn't supposed to be
super-fast (the function includes the 'msleep' call). The
function body is too big to be inline as well.